[ALPS07325121] wlan: add pcie interrupt debug log
1. add pcie status log
2. change atomic set/get for interrupt flag
3. enable debug sop interrupt

MTK-Commit-Id: cc0dea9927ac13caa5a1e332bc7d5f7232b94283

CR-Id: ALPS07325121
Feature: Wi-Fi Driver CONNAC
Reviewed-on: https://gerrit.mediatek.inc/c/neptune/wlan_driver/gen4m/+/6066569
CheckCourse: srv_pis <srv_pis@mediatek.com>
CheckCRInfo: srv_pis <srv_pis@mediatek.com>
CheckForbiddenKeyword: srv_pis <srv_pis@mediatek.com>
CheckPatch: srv_pis <srv_pis@mediatek.com>
Build: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Build: srv_preflight_a001 <srv_preflight_a001@mediatek.com>
Test: ch.ko <ch.ko@mediatek.com>

+++ include/nic/adapter.h
@@ -1580,7 +1580,7 @@ struct ADAPTER {
 
 	enum ENUM_ACPI_STATE rAcpiState;
 
-	u_int8_t fgIsIntEnable;
+	uint32_t fgIsIntEnable;
 	u_int8_t fgIsIntEnableWithLPOwnSet;
 
 	u_int8_t fgIsFwOwn;
